Troubleshooting moving ground management
The floors on World 3 like in scene W03_L01, have a moving ground that
on position reset randomzes mesh and soon collider mesh. They move
individualy in a line down the Z axis and when they get behind the
player they are moved far ahead of the player onto the end of the line.

MY PROBLEM

they are off on the Z axis my few fractions each time + or- revealing
the seams.
I have tried each of these with variation:
-destroy then initiate new down the z axis
-if self past -10, trans.pos down the z axis
-once initiated/trans.pos, Cast Ray to next floor collider and move gap
distance closer
-group controlled, if  a,b,c,d 's pos.z -10 then more down z

ive tried these ad more and they are all off just slightly.
So i am looking for a solution